Reuse test fixture function
